Restructured forms to accommodate L1A and L1B job forms. the /jobs/new endpoint now lists the available applications a job can be executed against, with endpoints for each application job execution form (e.g /jobs/new/chirp-rebinning).
