+++ src/routes/jobs/new/index.css
@@ -1,5 +1,26 @@
-form {
+.app-list-container {
    width: 600px;
    background-color: white;
    padding: 24px;
+}
+
+.app-list {
+   list-style-type: none;
+   padding: 0px;
+   margin: 0px;
+}
+
+.app-list li {
+   margin-bottom: 20px;
+}
+
+.app-list li:last-child {
+   margin-bottom: 0px;
+}
+
+.job-form {
+   background-color: white;
+   margin-top: 24px;
+   padding: 4px 24px 24px 24px;
+   width: 600px;
 }
